requirements.txt
streamlit
crewai
crewai-tools
langchain-openai
python-dotenv
yfinance
pandas
numpy